- Fix warnings on Solaris.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@29779 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/writers/writedb/writedb_isam.cpp
@@ -222,7 +222,7 @@ void CWriteDB_IsamIndex::x_WriteHeader()
     
     WriteInt4(isam_version);
     WriteInt4(isam_type);
-    WriteInt4(m_DataFileSize);
+    WriteInt4((int)m_DataFileSize);
     
     WriteInt4(num_terms);
     WriteInt4(samples);
@@ -258,7 +258,7 @@ void CWriteDB_IsamIndex::x_FlushStringIndex()
     // make reallocation rare, I divide by 63 instead, and throw in an
     // extra 16 bytes.
     
-    key_buffer.reserve(m_DataFileSize/63 + 16);
+    key_buffer.reserve((int)(m_DataFileSize/63 + 16));
     key_off.reserve(nsamples);
     
     unsigned i(0);
